requirements.txt
pillow
opencv_python
torchvision
torch
numpy
tqdm
matplotlib